Allow creating a sharable link
This allows saving the current dashboard to a shareable link that
contains a BASE64 encoded compressed JSON blob of the dashboard.

This also fixes an issue where the state from local storage was saved
instead of the current window state (this could result in saving the
wrong browser window if multiple windows were open).