Update NPM packages
Includes a fix for CVE-2022-31160 in jQuery UI. This vulnerability does
not seem to impact munin-plot because we don't use radio buttons.

Includes a fix for CVE-2022-24785 in Moment.js This vulnerability does
not seem to impact munin-plot because it should only affect server-side
Javascript.

Includes a fix for CVE-2022-25858 in terser. This is used by webpack and
should not affect impact munin-plot because it does not run webpack on
untrusted input.

Sadly no fix for CVE-2022-25758 is available at this time because there
is not yet a version of node-sass available that doesn't depend on the
scss-tokenizer package (which appears to be unmaintained). Since we
don't process untrusted SCSS it should not affect munin-plot.