Update NPM packages
Includes a fix for CVE-2022-25758 in scss-tokenizer. This vulnerability
does not seem to impact munin-plot because we don't run node-sass on
untrusted input.

Sadly no fix for a vulnerability in d3-color is available yet (no CVE
for GHSA-36jr-mh4h-2g58 has been assigned) because there is not yet a
version of plotly.js available that doesn't depend on a version of
d3-interpolate that doesn't depend on the vulnerable package. This could
affect munin-plot because d3 is unsed in the web application but we only
pass data to d3 (through plotly.js) that is output from the the
munin-plot server-side component.