Update NPM packages
This upgrades to version 7 of eslint which results in some minor other
changes.

Sadly no fix for CVE-2022-46175 in json5 (a transitive dependency of
eslint) is available yet. This vulnerability should not affect
munin-plot because we don't run eslint on untrusted input.

Sadly no fix for a vulnerability in d3-color is available yet (no CVE
for GHSA-36jr-mh4h-2g58 has been assigned) because there is not yet a
version of plotly.js available that doesn't depend on a version of
d3-interpolate that doesn't depend on the vulnerable package. This could
affect munin-plot because d3 is used in the web application but we only
pass data to d3 (via plotly.js) that is output of the munin-plot
server-side component.