Update NPM packages
Includes a fix for a vulnerability in d3-color is available yet (no CVE
for GHSA-36jr-mh4h-2g58 has been assigned). Since we only pass data to
d3 that is generated by the server-side component this should not affect
munin-plot.

Also includes a fix for CVE-2022-25881 in http-cache-semantics which
does not affect munin-plot because it only affects server-side
applications.