Update NPM packages
Includes a partial fix CVE-2023-26115 (Regular Expression Denial of
Service) in word-wrap. While word-wrap was used in the built application
the vulnerable version is now only used in a dependency of munin-plot
build tools. It should not have been run on any untrusted data.

This does not completely fix CVE-2022-25883 (Regular Expression Denial
of Service) in semver because of dependency issues. The semver package
is only used to build the resulting Javascript and as such is not run on
untrusted data.