Update NPM packages
This includes a fix for CVE-2024-21538 in cross-spawn (used by archiver,
eslint and webpack-cli). This vulnerability should not affect munin-plot
because none of the tools are run on untrusted data.

This also incldues a fix for CVE-2024-55565 in nanoid (used by
autoprefixer). This vulnerability should not affect munin-plot because
it is not called with untrusted data.